LPD-8529 Move methods from AssetRSSHelper to AssetPublisherPortlet and change references to be protected, since AssetPublisherPortlet is extended by others components and inline getRSS method into _getRSS
