+++ modules/apps/asset/asset-publisher-web/src/main/java/com/liferay/asset/publisher/web/internal/portlet/AssetPublisherPortlet.java
@@ -7,6 +7,7 @@ package com.liferay.asset.publisher.web.internal.portlet;
 
 import com.liferay.asset.constants.AssetWebKeys;
 import com.liferay.asset.kernel.AssetRendererFactoryRegistryUtil;
+import com.liferay.asset.kernel.model.AssetEntry;
 import com.liferay.asset.kernel.model.AssetRendererFactory;
 import com.liferay.asset.kernel.model.ClassType;
 import com.liferay.asset.kernel.model.ClassTypeField;
@@ -15,11 +16,13 @@ import com.liferay.asset.list.asset.entry.provider.AssetListAssetEntryProvider;
 import com.liferay.asset.list.service.AssetListEntrySegmentsEntryRelLocalService;
 import com.liferay.asset.publisher.constants.AssetPublisherPortletKeys;
 import com.liferay.asset.publisher.constants.AssetPublisherWebKeys;
+import com.liferay.asset.publisher.util.AssetEntryResult;
 import com.liferay.asset.publisher.util.AssetPublisherHelper;
+import com.liferay.asset.publisher.web.internal.configuration.AssetPublisherSelectionStyleConfigurationUtil;
 import com.liferay.asset.publisher.web.internal.configuration.AssetPublisherWebConfiguration;
+import com.liferay.asset.publisher.web.internal.constants.AssetPublisherSelectionStyleConstants;
 import com.liferay.asset.publisher.web.internal.display.context.AssetPublisherDisplayContext;
 import com.liferay.asset.publisher.web.internal.helper.AssetPublisherWebHelper;
-import com.liferay.asset.publisher.web.internal.helper.AssetRSSHelper;
 import com.liferay.asset.publisher.web.internal.util.AssetPublisherCustomizerRegistry;
 import com.liferay.asset.util.AssetHelper;
 import com.liferay.dynamic.data.mapping.model.DDMFormField;
@@ -32,12 +35,17 @@ import com.liferay.dynamic.data.mapping.util.DDMUtil;
 import com.liferay.fragment.processor.PortletRegistry;
 import com.liferay.info.item.InfoItemServiceRegistry;
 import com.liferay.item.selector.ItemSelector;
+import com.liferay.petra.string.StringBundler;
+import com.liferay.petra.string.StringPool;
 import com.liferay.portal.configuration.metatype.bnd.util.ConfigurableUtil;
+import com.liferay.portal.kernel.dao.search.SearchContainer;
 import com.liferay.portal.kernel.exception.NoSuchGroupException;
 import com.liferay.portal.kernel.json.JSONFactory;
 import com.liferay.portal.kernel.json.JSONObject;
+import com.liferay.portal.kernel.language.Language;
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
+import com.liferay.portal.kernel.model.Layout;
 import com.liferay.portal.kernel.model.Release;
 import com.liferay.portal.kernel.portlet.PortletIdCodec;
 import com.liferay.portal.kernel.portlet.bridges.mvc.MVCPortlet;
@@ -50,10 +58,19 @@ import com.liferay.portal.kernel.theme.ThemeDisplay;
 import com.liferay.portal.kernel.util.ContentTypes;
 import com.liferay.portal.kernel.util.DateFormatFactoryUtil;
 import com.liferay.portal.kernel.util.GetterUtil;
+import com.liferay.portal.kernel.util.Http;
 import com.liferay.portal.kernel.util.HttpComponentsUtil;
 import com.liferay.portal.kernel.util.ParamUtil;
 import com.liferay.portal.kernel.util.Portal;
+import com.liferay.portal.kernel.util.Validator;
 import com.liferay.portal.kernel.util.WebKeys;
+import com.liferay.rss.export.RSSExporter;
+import com.liferay.rss.model.SyndContent;
+import com.liferay.rss.model.SyndEntry;
+import com.liferay.rss.model.SyndFeed;
+import com.liferay.rss.model.SyndLink;
+import com.liferay.rss.model.SyndModelFactory;
+import com.liferay.rss.util.RSSUtil;
 import com.liferay.segments.SegmentsEntryRetriever;
 import com.liferay.segments.context.RequestContextMapper;
 
@@ -63,7 +80,9 @@ import java.io.Serializable;
 
 import java.text.DateFormat;
 
+import java.util.ArrayList;
 import java.util.Date;
+import java.util.List;
 import java.util.Map;
 
 import javax.portlet.ActionRequest;
@@ -71,6 +90,8 @@ import javax.portlet.ActionResponse;
 import javax.portlet.Portlet;
 import javax.portlet.PortletException;
 import javax.portlet.PortletPreferences;
+import javax.portlet.PortletRequest;
+import javax.portlet.PortletResponse;
 import javax.portlet.RenderRequest;
 import javax.portlet.RenderResponse;
 import javax.portlet.ResourceRequest;
@@ -290,9 +311,6 @@ public class AssetPublisherPortlet extends MVCPortlet {
 	protected AssetPublisherWebHelper assetPublisherWebHelper;
 
 	@Reference
-	protected AssetRSSHelper assetRSSHelper;
-
-	@Reference
 	protected InfoItemServiceRegistry infoItemServiceRegistry;
 
 	@Reference
@@ -300,6 +318,9 @@ public class AssetPublisherPortlet extends MVCPortlet {
 
 	@Reference
 	protected JSONFactory jsonFactory;
+
+	@Reference
+	protected Language language;
 
 	@Reference
 	protected Portal portal;
@@ -316,7 +337,168 @@ public class AssetPublisherPortlet extends MVCPortlet {
 	protected RequestContextMapper requestContextMapper;
 
 	@Reference
+	protected RSSExporter rssExporter;
+
+	@Reference
 	protected SegmentsEntryRetriever segmentsEntryRetriever;
+
+	@Reference
+	protected SyndModelFactory syndModelFactory;
+
+	private String _exportToRSS(
+			PortletRequest portletRequest, PortletResponse portletResponse,
+			String name, String format, double version, String displayStyle,
+			String linkBehavior, List<AssetEntry> assetEntries)
+		throws Exception {
+
+		ThemeDisplay themeDisplay = (ThemeDisplay)portletRequest.getAttribute(
+			WebKeys.THEME_DISPLAY);
+
+		SyndFeed syndFeed = syndModelFactory.createSyndFeed();
+
+		syndFeed.setDescription(name);
+
+		List<SyndEntry> syndEntries = new ArrayList<>();
+
+		syndFeed.setEntries(syndEntries);
+
+		for (AssetEntry assetEntry : assetEntries) {
+			SyndEntry syndEntry = syndModelFactory.createSyndEntry();
+
+			syndEntry.setAuthor(portal.getUserName(assetEntry));
+
+			SyndContent syndContent = syndModelFactory.createSyndContent();
+
+			syndContent.setType(RSSUtil.ENTRY_TYPE_DEFAULT);
+
+			String value = null;
+
+			String languageId = language.getLanguageId(portletRequest);
+
+			if (displayStyle.equals(RSSUtil.DISPLAY_STYLE_TITLE)) {
+				value = StringPool.BLANK;
+			}
+			else {
+				value = assetEntry.getSummary(languageId, true);
+			}
+
+			syndContent.setValue(value);
+
+			syndEntry.setDescription(syndContent);
+
+			String link = _getEntryURL(
+				portletRequest, portletResponse, linkBehavior, assetEntry);
+
+			syndEntry.setLink(link);
+
+			syndEntry.setPublishedDate(assetEntry.getPublishDate());
+			syndEntry.setTitle(assetEntry.getTitle(languageId, true));
+			syndEntry.setUpdatedDate(assetEntry.getModifiedDate());
+			syndEntry.setUri(link);
+
+			syndEntries.add(syndEntry);
+		}
+
+		syndFeed.setFeedType(RSSUtil.getFeedType(format, version));
+
+		List<SyndLink> syndLinks = new ArrayList<>();
+
+		syndFeed.setLinks(syndLinks);
+
+		SyndLink selfSyndLink = syndModelFactory.createSyndLink();
+
+		syndLinks.add(selfSyndLink);
+
+		String feedURL = _getFeedURL(portletRequest);
+
+		selfSyndLink.setHref(feedURL);
+
+		selfSyndLink.setRel("self");
+
+		SyndLink alternateSyndLink = syndModelFactory.createSyndLink();
+
+		syndLinks.add(alternateSyndLink);
+
+		alternateSyndLink.setHref(portal.getLayoutFullURL(themeDisplay));
+		alternateSyndLink.setRel("alternate");
+
+		syndFeed.setPublishedDate(new Date());
+		syndFeed.setTitle(name);
+		syndFeed.setUri(feedURL);
+
+		return rssExporter.export(syndFeed);
+	}
+
+	private List<AssetEntry> _getAssetEntries(
+			PortletRequest portletRequest,
+			PortletPreferences portletPreferences)
+		throws Exception {
+
+		ThemeDisplay themeDisplay = (ThemeDisplay)portletRequest.getAttribute(
+			WebKeys.THEME_DISPLAY);
+
+		List<AssetEntry> assetEntries = new ArrayList<>();
+
+		SearchContainer<AssetEntry> searchContainer = new SearchContainer();
+
+		AssetPublisherDisplayContext assetPublisherDisplayContext =
+			(AssetPublisherDisplayContext)portletRequest.getAttribute(
+				AssetPublisherWebKeys.ASSET_PUBLISHER_DISPLAY_CONTEXT);
+
+		searchContainer.setDelta(assetPublisherDisplayContext.getRSSDelta());
+
+		Map<String, Serializable> attributes =
+			assetPublisherDisplayContext.getAttributes();
+
+		attributes.put("filterExpired", Boolean.TRUE);
+
+		List<AssetEntryResult> assetEntryResults =
+			assetPublisherHelper.getAssetEntryResults(
+				searchContainer,
+				assetPublisherDisplayContext.getAssetEntryQuery(),
+				themeDisplay.getLayout(), portletPreferences,
+				assetPublisherDisplayContext.getPortletName(),
+				themeDisplay.getLocale(), themeDisplay.getTimeZone(),
+				themeDisplay.getCompanyId(), themeDisplay.getScopeGroupId(),
+				themeDisplay.getUserId(),
+				assetPublisherDisplayContext.getClassNameIds(), attributes);
+
+		for (AssetEntryResult assetEntryResult : assetEntryResults) {
+			assetEntries.addAll(assetEntryResult.getAssetEntries());
+		}
+
+		return assetEntries;
+	}
+
+	private String _getAssetPublisherURL(PortletRequest portletRequest)
+		throws Exception {
+
+		ThemeDisplay themeDisplay = (ThemeDisplay)portletRequest.getAttribute(
+			WebKeys.THEME_DISPLAY);
+
+		Layout layout = themeDisplay.getLayout();
+
+		PortletDisplay portletDisplay = themeDisplay.getPortletDisplay();
+
+		StringBundler sb = new StringBundler(6);
+
+		String layoutFriendlyURL = GetterUtil.getString(
+			portal.getLayoutFriendlyURL(layout, themeDisplay));
+
+		if (!layoutFriendlyURL.startsWith(Http.HTTP_WITH_SLASH) &&
+			!layoutFriendlyURL.startsWith(Http.HTTPS_WITH_SLASH)) {
+
+			sb.append(themeDisplay.getPortalURL());
+		}
+
+		sb.append(layoutFriendlyURL);
+		sb.append(Portal.FRIENDLY_URL_SEPARATOR);
+		sb.append("asset_publisher/");
+		sb.append(portletDisplay.getInstanceId());
+		sb.append(StringPool.SLASH);
+
+		return sb.toString();
+	}
 
 	private String _getDisplayFieldValue(Field field, ThemeDisplay themeDisplay)
 		throws Exception {
@@ -342,6 +524,61 @@ public class AssetPublisherPortlet extends MVCPortlet {
 		}
 
 		return fieldValue;
+	}
+
+	private String _getEntryURL(
+			PortletRequest portletRequest, PortletResponse portletResponse,
+			String linkBehavior, AssetEntry assetEntry)
+		throws Exception {
+
+		if (linkBehavior.equals("viewInPortlet")) {
+			return _getEntryURLViewInContext(
+				portletRequest, portletResponse, assetEntry);
+		}
+
+		return _getEntryURLAssetPublisher(portletRequest, assetEntry);
+	}
+
+	private String _getEntryURLAssetPublisher(
+			PortletRequest portletRequest, AssetEntry assetEntry)
+		throws Exception {
+
+		AssetRendererFactory<?> assetRendererFactory =
+			AssetRendererFactoryRegistryUtil.getAssetRendererFactoryByClassName(
+				assetEntry.getClassName());
+
+		return StringBundler.concat(
+			_getAssetPublisherURL(portletRequest),
+			assetRendererFactory.getType(), "/id/", assetEntry.getEntryId());
+	}
+
+	private String _getEntryURLViewInContext(
+		PortletRequest portletRequest, PortletResponse portletResponse,
+		AssetEntry assetEntry) {
+
+		String assetViewURL = assetPublisherHelper.getAssetViewURL(
+			portal.getLiferayPortletRequest(portletRequest),
+			portal.getLiferayPortletResponse(portletResponse), assetEntry,
+			true);
+
+		if (Validator.isNotNull(assetViewURL) &&
+			!assetViewURL.startsWith(Http.HTTP_WITH_SLASH) &&
+			!assetViewURL.startsWith(Http.HTTPS_WITH_SLASH)) {
+
+			ThemeDisplay themeDisplay =
+				(ThemeDisplay)portletRequest.getAttribute(
+					WebKeys.THEME_DISPLAY);
+
+			assetViewURL = themeDisplay.getPortalURL() + assetViewURL;
+		}
+
+		return assetViewURL;
+	}
+
+	private String _getFeedURL(PortletRequest portletRequest) throws Exception {
+		String feedURL = _getAssetPublisherURL(portletRequest);
+
+		return feedURL.concat("rss");
 	}
 
 	private void _getFieldValue(
@@ -513,10 +750,38 @@ public class AssetPublisherPortlet extends MVCPortlet {
 				AssetPublisherWebKeys.ASSET_PUBLISHER_DISPLAY_CONTEXT,
 				assetPublisherDisplayContext);
 
-			byte[] bytes = assetRSSHelper.getRSS(
-				resourceRequest, resourceResponse);
+			portletPreferences = resourceRequest.getPreferences();
 
-			outputStream.write(bytes);
+			String selectionStyle = portletPreferences.getValue(
+				"selectionStyle",
+				AssetPublisherSelectionStyleConfigurationUtil.
+					defaultSelectionStyle());
+
+			if (!selectionStyle.equals(
+					AssetPublisherSelectionStyleConstants.TYPE_DYNAMIC)) {
+
+				outputStream.write(new byte[0]);
+
+				return;
+			}
+
+			String assetLinkBehavior = portletPreferences.getValue(
+				"assetLinkBehavior", "showFullContent");
+			String rssDisplayStyle = portletPreferences.getValue(
+				"rssDisplayStyle", RSSUtil.DISPLAY_STYLE_ABSTRACT);
+			String rssFeedType = portletPreferences.getValue(
+				"rssFeedType", RSSUtil.FEED_TYPE_DEFAULT);
+			String rssName = portletPreferences.getValue("rssName", null);
+
+			String format = RSSUtil.getFeedTypeFormat(rssFeedType);
+			double version = RSSUtil.getFeedTypeVersion(rssFeedType);
+
+			String rss = _exportToRSS(
+				resourceRequest, resourceResponse, rssName, format, version,
+				rssDisplayStyle, assetLinkBehavior,
+				_getAssetEntries(resourceRequest, portletPreferences));
+
+			outputStream.write(rss.getBytes(StringPool.UTF8));
 		}
 		catch (Exception exception) {
 			_log.error("Unable to get RSS feed", exception);
